NLP_translator_web_app/requirements.txt
googletrans==3.1.0a0
gTTS==2.3.2
ipython==8.15.0
Pillow==9.5.0
Pillow==10.0.0
playsound==1.2.2
SpeechRecognition==3.10.0
streamlit==1.22.0
